% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_key_free2-data.R
\docType{data}
\name{answer_key_free2}
\alias{answer_key_free2}
\title{Answer Key Example Data}
\format{
A data frame of answers for a free recall test

Answer_Key: a list of free recall answers
}
\usage{
data(answer_key_free2)
}
\description{
Dataset that includes the answer key for free recall data.
Pair with the free_data dataset for examples.
}
\keyword{datasets}
