% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data.R
\name{arrange.data}
\alias{arrange.data}
\title{Arrange Data for Free Recall Scoring}
\usage{
\method{arrange}{data}(x, sep = y, id = z, other = NULL)
}
\arguments{
\item{x}{a vector containing participant responses}

\item{sep}{a character separating each response. This needs to be in quotes.}

\item{id}{a vector containing participant ID numbers}

\item{Other}{an optional argument for passing condition columns. Needs to take the form of an index of the original dataframe.}
}
\value{
a dataframe object containing the input vectors and a percent match column
}
\description{
This function takes wide format free recall data where all responses are stored in the same cell and converts it to long format.
}
