% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{balOneSH}
\alias{balOneSH}
\title{Univariate covariance-adjusted balance test, using Huber-White SE}
\usage{
balOneSH(dat, BW, varb, rhs = NULL)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R} and a covariate}

\item{BW}{The bandwidth within which to test balance.}

\item{varb}{The character name of the variable to test balance on.}

\item{rhs}{A string specifying the \eqn{\mu_\beta(\cdot)} model relating the outcome to
\code{R}. Defaults to \code{"R+Z"}}
}
\value{
scalar, the p-value associated w/ coefficient on Z
}
\description{
The test is the test associated with the Z coefficient after fitting
a linear or logistic-linear model with Z and R as independent variables.
The Huber-White SE is the one provided by the sandwich package.
#'
}
