% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsandwich.R
\name{bread.lmrob}
\alias{bread.lmrob}
\title{Bread method for lmrob objects}
\usage{
\method{bread}{lmrob}(x, ...)
}
\arguments{
\item{x}{an lmrob object produced using an MM/SM estimator chain}

\item{...}{}
}
\value{
k by (k+1) matrix, with first column for scale estimate and rows, remaining cols for coefficients
}
\description{
Extract bread matrix from an lmrob fit
}
\details{
This is part of a workaround for an issue in the robustbase code
affecting sandwich covariance estimation.
The issue in question is issue #6471, robustbase project on R-Forge.
}
\author{
lrd author 2
}
