% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bw}
\alias{bw}
\title{Choose a bandwidth using either Limitless or Permutations method with one covariate, or the IK method}
\usage{
bw(dat, covname = "x", method = "sh", alphax = 0.15, plt = FALSE)
}
\arguments{
\item{dat}{a data.frame. must contain variables `R` (the running variable) and `Z`
(treatment) and, if method='ik', `Y` (outcome)}

\item{covname}{a character string, the name of the covariate to test balance on
(unless method='ik')}

\item{method}{a character string. either 'sh' for Limitless, 'ik' for IK, or 'cft' for
Permutations}
}
\value{
bandwidth choice (scalar)
}
\description{
Choose a bandwidth using either Limitless or Permutations method with one covariate, or the IK method
}
