% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bwMult}
\alias{bwMult}
\title{Choose a bandwidth using one of several methods}
\usage{
bwMult(dat, alpha = 0.15, balMult.control = list(method = "sh",
  reduced.covars = FALSE), bws = seq(0, 2, 0.01))
}
\arguments{
\item{dat}{a data.frame. must contain variables `R` (the running variable) and `Z`
(treatment) and, if method='ik', `Y` (outcome) along with these covariates: 'lhsgrade_pct',
'totcredits_year1','male','loc_campus1','loc_campus2','bpl_north_america','english',
and 'age'}

\item{alpha}{the level of the balance test}

\item{newbal.control}{list of optional arguments to `newBal`}

\item{bsw}{numeric, a sequence of bandwidths to try (in increasing order)}
}
\value{
bandwidth choice (scalar)
}
\description{
Method options include Limitless, a permutation-testing method, or the IK method
}
