% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cft}
\alias{cft}
\title{CFT analysis method}
\usage{
cft(dat, BW, outcome, alpha = 0.05, rhs = NULL)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{(Optional) A bandwidth \code{b>0}. If not provided it will be estimated from the data}

\item{outcome}{A string specifying the name of the outcome or covariate variable}

\item{alpha}{1-confidence interval level}

\item{rhs}{Ignored}
}
\value{
list of
\describe{
 \item{p.value} of Fisher-style no effect hypothesis
 \item{CI} vector of confidence limits, estimate
 \item{bal.pval} balance p-value
 \item{n} number of observations inselected window
 \item{W}  window
}
}
\description{
RDD analysis method modeled on Cattaneo et al 2014, J Causal Inference
}
\author{
lrd author 1
}
