% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{CIsh}
\alias{CIsh}
\title{Confidence Interval and Hodges-Lehmann estimate of a treatment effect for limitless
regression discontinuity}
\usage{
CIsh(dat, BW, outcome = "Y", est, alpha = 0.05, rhs = rhs)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{A bandwidth \code{b>0}. The test uses data for subjects with |R|<b}

\item{outcome}{A string specifying the name of the outcome variable}

\item{est}{A point estimate generated by \code{HLsh}. If not provided it will be estimated
automatically.}

\item{alpha}{the level of the confidence interval.}

\item{rhs}{A string specifying the \eqn{\mu_\beta(\cdot)} model relating the outcome to \code{R}. Defaults to \code{"R+Z"}}
}
\value{
a vector including
\describe{
 \item{CI1} and
 \item{CI2}, the lower and upper bounds of the confidence interval;
 \item{est}, the HL point estimate
}
}
\description{
Inverts the test in \code{testSH()} to find the Confidence Interval and Hodges-Lehmann
point estimate
}
