% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cued_data.R
\docType{data}
\name{cued_data}
\alias{cued_data}
\title{Cued Recall Data}
\format{
A data frame of answers for a cued recall test data

id: the participant id
trial: the trial id
response: the response the participant gave to the cue
key: the answer for this trial id
condition: the between subjects group the participants were in
}
\usage{
data(cued_data)
}
\description{
Dataset that includes cued recall data in long format.
Participants were given a cue, and they were required to
remember the response listed in the dataset. This dataset
is in long format, which is required for most functions.
}
\keyword{datasets}
