% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cued_data_groupby.R
\docType{data}
\name{cued_data_groupby}
\alias{cued_data_groupby}
\title{Cued Recall Data with Multiple Conditions}
\format{
A data frame of answers for a cued recall test data

Subject: the participant id
Target: the answer for this trial id
Response: the response the participant gave to the cue
Condition: the between subjects group the participants
 were in
Condition2: the second between subjects group the
 participants were in
}
\usage{
data(cued_data_groupby)
}
\description{
Dataset that includes cued recall data in long format.
Participants were given a cue, and they were required to
remember the response listed in the dataset. This dataset
is in long format, which is required for most functions.
}
\keyword{datasets}
