% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cued_recall_manuscript-data.R
\docType{data}
\name{cued_recall_manuscript}
\alias{cued_recall_manuscript}
\title{Cued Recall Data from Manuscript}
\format{
A data frame of answers for a cued recall test data

Sub.ID: the participant id
Trial_num: the trial id
Cue: the cue shown to participants
Target: the answer for this trial id
Answer: the participant answer for this trial
}
\usage{
data(cued_data)
}
\description{
Dataset that includes cued recall data in long format.
Participants were given a cue, and they were required to
remember the response listed in the dataset. This dataset
is in long format, which is required for most functions.
}
\keyword{datasets}
