% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_condition.R
\name{describe.condition}
\alias{describe.condition}
\title{Describe data}
\usage{
\method{describe}{condition}(x, group.by = y, id = z)
}
\arguments{
\item{x}{a vector containing input data}
}
\value{
prints descriptive statistics
}
\description{
This function provides descriptive statistics for a dataset, grouped by a condition
}
