% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsandwich.R
\name{estfun.lmrob}
\alias{estfun.lmrob}
\title{Estfun method for lmrob objects}
\usage{
\method{estfun}{lmrob}(x, ...)
}
\arguments{
\item{x}{a fitted lmrob}

\item{...}{}
}
\value{
an estfun object, as in the sandwich package
}
\description{
Only SM or MM estimates supported
}
\author{
lrd author 2
}
