% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetchLSOdata}
\alias{fetchLSOdata}
\title{Fetch Lindo et al data}
\usage{
fetchLSOdata(whereto = "extdata")
}
\arguments{
\item{whereto}{path specification for directory to data file to}
}
\value{
path of downloaded dta file
}
\description{
Retrieve Lindo et al replication materials, including data
}
\details{
Downloads the data from the AEJ website.

The \code{whereto} parameter should be a character string not ending
in \sQuote{\code{/}} (the path separator).
}
\author{
Ben B Hansen
}
