% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{frandsen}
\alias{frandsen}
\title{Frandsen Manipulation Test}
\usage{
frandsen(R, cutoff, BW, k = c(0, 0.01, 0.02, 0.1))
}
\arguments{
\item{R}{a discrete running variable}

\item{cutoff}{the value of the running variable at the boundary between treatment and control---in Frandsen (2016), the maximum (minimum) R at which subjects are assigned to treatment; alternatively, the minimum (maximum) R at which subjects are assigned to control.}

\item{BW}{an optional bandwidth, helpful for choosing a benchmark for k}

\item{k}{a vector of possible values of k, a tuning parameter for the test. The function automatically computes the benchmark value Frandsen (2016) suggests based on normal theory, and adds it to this list.}
}
\value{
a p-value
}
\description{
Frandsen Manipulation Test
}
