% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_data.R
\docType{data}
\name{free_data}
\alias{free_data}
\title{Free Recall Data}
\format{
A data frame of answers for a free recall test data

Username: the participant id
List_Types: a repeated measures condition participants were in
Response: the response the participant gave to the cue
Version: the version of the list_type given
Batch: the batch of participants that were run together
}
\usage{
data(free_data)
}
\description{
Dataset that includes free recall data in long format.
Participants were given a list of words to remember, and
then asked to recall the words. This dataset
is in wide format, which should be converted with arrange
data.
}
\keyword{datasets}
