% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_dataset}
\alias{get_dataset}
\title{Get Dataset}
\usage{
get_dataset(
  corpus = NULL,
  bibtexID = NULL,
  citation = NULL,
  language = NULL,
  variables = NULL
)
}
\arguments{
\item{corpus}{Include a two letter code to download the Open Subtitles corpus for
text models. You can view the corpora on
\href{http://opus.nlpl.eu/OpenSubtitles.php}{their website}. Note: these files
can be very large, so they may take up a lot of memory to download. They are
text based files that are read using `readLines`.}

\item{bibtexID}{The bibtex ID of the dataset you are trying to load.
You can leave all parameters blank to load just the metadata.}

\item{citation}{Include the citation for the dataset you loaded - will only
load if you include a bibtex ID.}

\item{language}{If you include a bibtex ID, you will get back the language of
the dataset, if you do not include a bibtex ID, it will return a list of
datasets in that language.}

\item{variables}{If you include a bibtex ID, you will get back the variables
included the dataset, if you do not include a bibtex ID, it will return a list of
datasets that include that variable (can also be paired with language).
Use the column names from the metadata as your filter.}
}
\value{
\item{metadata}{The metadata list of avaliable datasets}
\item{loaded_data}{The dataset you requested to load}
\item{language}{The language of the dataset you requested to load}
\item{variables}{The variables of the dataset you requested to load}
\item{datasets}{Possible datasets based on your language and variable names}
}
\description{
This function allows you to import the current datasets avaliable from
\href{https://github.com/orgs/SemanticPriming/}{Semantic Priming GitHub Group}.
}
\examples{

get_dataset()
get_dataset(bibtexID = "Birchenough2017", citation = TRUE)
get_dataset(language = "English", variables = c("aoa", "freq"))
}
\keyword{datasets,}
\keyword{linguistic}
\keyword{metadata,}
\keyword{norms}
