% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{HLsh}
\alias{HLsh}
\title{Hodges-Lehmann estimate of a treatment effect for limitless regression discontinuity}
\usage{
HLsh(dat, BW, outcome = "Y", rhs = NULL)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{A bandwidth \code{b>0}. The test uses data for subjects with |R|<b}

\item{outcome}{A string specifying the name of the outcome variable}

\item{rhs}{A string specifying the \eqn{\mu_\beta(\cdot)} model relating the outcome to
\code{R}. Defaults to \code{"R+Z"}}
}
\value{
scalar, the Hodges-Lehmann estimate
}
\description{
Inverts the test in \code{testSH()} to find the Hodges-Lehmann point estimate
}
