% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ik}
\alias{ik}
\title{Wrapper for Imbens-Kalyanaraman style RDD analysis}
\usage{
ik(dat, BW = NULL, outcome)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{(Optional) A bandwidth \code{b>0}. If not provided it will be estimated from the
data}

\item{outcome}{A string specifying the name of the outcome variable}
}
\value{
a list consisting of
\describe{
 \item{p.value} The p-value testing no effect
 \item{CI} a vector of confidence limits and the HL treatment effect
 \item{BW} the RDD bandwidth
 \item{bal.pval} a p-value testing for covariate balance
 \item{n} the number of subjects in the window of analysis
 \item{W} the range of R values in the window of analysis
}
}
\description{
Uses OLS to estimate the RDD Local Average Treatment Effect (LATE) as in
Imbens, G. and Kalyanaraman, K. (2012). Uses a "rectangular" kernal.
Tests balance on a set of pre-treatment covariates using \code{balMult} and either the
provided bandwidth or the bandwidth calculated with the IK procedure.
}
