% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{ikSim}
\alias{ikSim}
\title{Bandwidth p-value, point estimate from IK method}
\usage{
ikSim(dat, BW = NULL)
}
\arguments{
\item{dat}{data frame with variables R, x, Y, Z}
}
\value{
vector with elements
\itemize{
 \item bandwidth
 \item p-value
 \item HL point estimate
}
}
\description{
Analyze a simulated dataset with the IK method (implemented via rdd package)
}
