% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ikTest}
\alias{ikTest}
\title{Imbens-Kalyanaraman style RDD analysis}
\usage{
ikTest(dat, BW = NULL, varb, rhs = NULL, justP = TRUE)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{(Optional) A bandwidth \code{b>0}. If not provided it will be
estimated from the data}

\item{varb}{A string specifying the name of the outcome or covariate variable}

\item{rhs}{Ignored}

\item{justP}{if TRUE only returns the p-value}
}
\value{
if justP=TRUE, returns the p-value for an effect. if justP=FALSE,
returns the fitted RDD model (an object of class "RD").
}
\description{
Uses OLS to estimate the RDD Local Average Treatment Effect (LATE) as in
Imbens, G. and Kalyanaraman, K. (2012). Uses a "rectangular" kernal.
Tests balance on a set of pre-treatment covariates using \code{balMult} and either the
provided bandwidth or the bandwidth calculated with the IK procedure.
}
