% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{kappa}
\alias{kappa}
\title{Cohen's Kappa}
\usage{
kappa(x, y)
}
\arguments{
\item{x}{rater 1}

\item{y}{rater 2}
}
\value{
Cohen's Kappa value
}
\description{
This function returns Cohen's Kappa for two raters
}
