% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{makeData}
\alias{makeData}
\title{Create simulated data set}
\usage{
makeData(n, curve, tdist = FALSE, tau = 0)
}
\arguments{
\item{n}{size (no. obs) in simulation}

\item{curve}{Amount by which magnitude of slope increases as R passes below -.5}

\item{tdist}{Generate disturbances from t dist'n on 3 d.f.? (As opposed to std Normal,
the default)}

\item{tau}{Magnitude of simulated treatment effect}
}
\value{
Data frame w/ running variable R, covariate x, yc, Z, Y
}
\description{
With `curve=3` and `tdist=TRUE` we get the simulation as reported in arxiv v3
}
