% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_answers-data.R
\docType{data}
\name{multi_answers}
\alias{multi_answers}
\title{Answer Key Example Data for Multiple Lists}
\format{
A data frame of answers for a free recall test

List1: a list of free recall answers
List2: a second list of free recall answers
etc.
}
\usage{
data(multi_answers)
}
\description{
Dataset that includes the answer key for free recall data.
Pair with the multi_data dataset for examples.
}
\keyword{datasets}
