% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_data-data.R
\docType{data}
\name{multi_data}
\alias{multi_data}
\title{Free Recall Data in Wide Format with Multiple Lists}
\format{
A data frame of answers for a free recall test data

Sub.ID: the participant id
List.Type: the type of list a person saw
Response: the response the participant gave to the cue
List.Number: the number of the list they completed
}
\usage{
data(multi_data)
}
\description{
Dataset that includes free recall data in long format.
Participants were given a list of words to remember, and
then asked to recall the words. This dataset
is in wide format, which should be converted with arrange
data.
}
\keyword{datasets}
