% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_agree.R
\name{percent_agree}
\alias{percent_agree}
\title{Percent Agree}
\usage{
percent_agree(x, y)
}
\arguments{
\item{x}{rater 1}

\item{y}{rater 2}
}
\description{
This function allows you to check the percentage that
that two raters agree in their ratings.
}
\examples{
percent_agree()
}
\keyword{percent}
