% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_percentages.R
\name{percent_match.cr}
\alias{percent_match.cr}
\title{Compute Percent Match cued-recall}
\usage{
percent_match.cr(x, key = y, id = z, weight = FALSE, weight.by = NULL)
}
\arguments{
\item{x}{a vector containing input data}

\item{key}{a vector containing the scoring key}

\item{id}{a vector containing participant ID numbers}
}
\value{
a dataframe object containing the input vectors and a percent match column
}
\description{
This function computes the percentage of shared characters between two strings
}
