% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight percentages.R
\name{percent_match}
\alias{percent_match}
\title{Compute Percent Match}
\usage{
percent_match(x, key = y, id = z, weight = FALSE, weight.by = NULL)
}
\arguments{
\item{x}{a vector containing input data}
}
\value{
a dataframe object containing the input vectors and a percent match column
}
\description{
This function computes the percentage of shared characters between two strings
}
