% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_match_s.R
\name{percent_match.w}
\alias{percent_match.w}
\title{Compute the Percent Match between two words}
\usage{
percent_match.w(x, y)
}
\arguments{
\item{x}{word 1, a character string}

\item{y}{word 2, a character string}
}
\value{
a percent match value
}
\description{
This function computes the percentage of shared characters between two strings
}
