% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion correct free recall.R
\name{prop.correct.f}
\alias{prop.correct.f}
\title{Proportion Correct Free Recalll}
\usage{
prop.correct.f(x, key = y, id = z, flag = FALSE, group.by = NULL)
}
\arguments{
\item{x}{a vector containing participant scores}

\item{key}{a vector containing the scoring key}

\item{id}{a vector containing participant ID numbers}

\item{group.by}{an optional argument that can be used to group the output by one condition column}
}
\value{
prints descriptive statistics. Output can be saved to a dataframe.
}
\description{
This function computes the proportion of correct responses per participant. Proportions can either be separated by condition or collapsed across conditions. Note: This function returns mean values when used with non-binary data.
}
