% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop_Correct.R
\name{prop.correct}
\alias{prop.correct}
\title{Proportion Correct}
\usage{
prop.correct(x, group.by = NULL, id = z, flag = FALSE)
}
\arguments{
\item{x}{a vector containing participant scores}
}
\value{
prints descriptive statistics. Output can be saved to a dataframe.
}
\description{
This function computes the proportion of correct responses per participant. Proportions can either be separated by condition or collapsed across conditions. Note: This function returns mean values when used with non-binary data.
}
