% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_data-data.R
\docType{data}
\name{rater_data}
\alias{rater_data}
\title{Rater Data}
\format{
A data frame of scored answers for inter-rater reliability

Sub.ID: the participant id
rater1_word: the word choice for the subject the rater selected
rater1_score: the score for the participant given by the rater
rater2_word: the word choice for the subject the rater selected
rater2_score: the score for the participant given by the rater
}
\usage{
data(rater_data)
}
\description{
Dataset that contains scoring and ratings for a recall test
that was rated by two raters. Use with the kappa function
as an example.
}
\keyword{datasets}
