% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsandwich.R
\name{sandwich}
\alias{sandwich}
\title{Sandwich estimate of covariance}
\usage{
sandwich(x, bread. = sandwich::bread, meat. = sandwich::meat, ...)
}
\arguments{
\item{x}{a fitted model object, as in sandwich::sandwich}

\item{bread.}{function or matrix,}

\item{meat.}{function or matrix, as in sandwich::sandwich}

\item{...}{additional arguments to downstream methods, as in sandwich::sandwich}
}
\value{
matrix, bread %*% meat %*% t(bread)
}
\description{
Overloading of sandwich::sandwich to accommodate non-square bread
}
\details{
The sandwich package's sandwich function presumes the bread matrix
to be symmetric. Obviously this won't do if the bread is rectangular but not
square.

This is part of a workaround for an issue in the robustbase code
affecting sandwich covariance estimation.
The issue in question is issue #6471, robustbase project on R-Forge.
}
\author{
lrd author 2
}
