% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score free recall.R
\name{score.recall.f}
\alias{score.recall.f}
\title{Score free recall data}
\usage{
score.recall.f(x, key = y, id = z, cutoff = g)
}
\arguments{
\item{x}{a dataframe object containing the scored data}

\item{cutoff}{a value used to specify the scoring cutoff. Must be a value between 0 and 1.}
}
\value{
a .csv file containing the scored output
}
\description{
Use this function to score free data based on percent match
}
