% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_function.R
\name{score_recall.cr}
\alias{score_recall.cr}
\title{Score cued-recall data}
\usage{
score_recall.cr(x, weight = FALSE, cutoff = y)
}
\arguments{
\item{x}{a dataframe object containing the scored data}

\item{set.cutoff}{a value used to specify the scoring cutoff. Must be a value between 0 and 1.}
}
\value{
a .csv file containing the scored output
}
\description{
Use this function to score cued-recall data based on percent match
}
