% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_function.R
\name{score_recall}
\alias{score_recall}
\title{Score Lexical Data}
\usage{
score_recall(x, weight = FALSE, set.cutoff = y)
}
\arguments{
\item{x}{a dataframe object containing the scored data}
}
\value{
a .csv file containing the scored output
}
\description{
Use this function to score lexical data based on percent match
}
