% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_data.R
\docType{data}
\name{sentence_data}
\alias{sentence_data}
\title{Sentence Recall Data}
\format{
A data frame of answers for a sentence recall test data

Sub.ID: the participant id
Trial.ID: the id for the trial given to participant
Sentence: the answer to the trial that the participant
 should have given
Response: the response the participant gave to that trial
Condition: the between subjects condition the participant
 was in
}
\usage{
data(sentence_data)
}
\description{
Dataset that includes sentence recall data in long format.
Participants were given a sentence to remember, and
then asked to recall the words. This dataset
is in long format, which is required for these functions.
}
\keyword{datasets}
