% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score Sentences.R
\name{sentence_match}
\alias{sentence_match}
\title{Sentence Match}
\usage{
sentence_match(x, key = y, id = z, cutoff = q)
}
\arguments{
\item{x}{A vector containing participant responses}

\item{key}{A vector containing the scoring key}

\item{id}{A vector containing participant ID numbers}

\item{cutoff}{a numerical value used for computing the percent match used at scoring. Must be a value between 0 and 1.}
}
\description{
This function allows you to compute the number of matching words between two sentences stored in a dataframe.
that two raters agree in their ratings.
}
\keyword{percent}
