% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{sh}
\alias{sh}
\title{Wrapper function for limitless regression discontinuity testing and estimation}
\usage{
sh(dat, BW, outcome, Dvar = "probation_year1", alpha = 0.05, rhs = NULL)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, an outcome variable, and (optionally)
an indicator for treatment received, in the case of partial compliance.}

\item{BW}{An optional bandwidth \code{b>0}. The test uses data for subjects with |R|<b. When
\code{BW} is omitted, the function uses sequential covariate balance tests to choose a
data-driven bandwidth.}

\item{outcome}{A string specifying the name of the outcome variable}

\item{Dvar}{An optional string specifying the name of the treatment-received variable in the
case of partial compliance.}

\item{alpha}{The level of the confidence interval}

\item{rhs}{A string specifying the \eqn{\mu_\beta(\cdot)} model relating the outcome to
\code{R}. Defaults to \code{"R+Z"}}
}
\value{
a list consisting of
\describe{
 \item{p.value} The p-value testing no effect
 \item{CI} a vector of confidence limits and the HL treatment effect
 \item{BW} the RDD bandwidth
 \item{bal.pval} a p-value testing for covariate balance
 \item{n} the number of subjects in the window of analysis
 \item{W} the range of R values in the window of analysis
}
}
\description{
In a regression discontinuity design, this function tests for an effect using \code{testSH},
and inverts the test to estimate a confidence interval and Hodges-Lehmann point estimate.
The test is the test associated with the Z coefficient after fitting
a model from the robustbase suite that uses Z and R as independent variables.
Covariance adjustment is via robust logistic regression if
ytilde is binary, robust linear regression otherwise. In the latter case,
since we're using `robustbase::lmrob` with the `MM` method, the SE used
in the test is of the Huber-White type, allowing for heteroskedasticity
and accounting for  propagation of errors from the preliminary fitting
done by the robust regression routine.
}
\details{
This is the method recommended in version 2 of Limitless
Regression Discontinuity
}
