% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{shbw}
\alias{shbw}
\title{Bandwidth, p-value, point estimate from SH method}
\usage{
shbw(ddd, BW = NULL)
}
\arguments{
\item{dat}{data frame with variables R, x, Y, Z}
}
\value{
vector with elements
\itemize{
 \item bandwidth
 \item p-value
 \item HL point estimate
}
}
\description{
Analyze a simulated data set with the SH method
}
