% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{testSH}
\alias{testSH}
\title{Robust balance test following robust covariate adjustment}
\usage{
testSH(dat, BW, tau = 0, outcome = "Y", return.coef = FALSE, rhs = NULL)
}
\arguments{
\item{dat}{Data set with columns \code{Z}, \code{R}, and an outcome variable}

\item{BW}{A bandwidth \code{b>0}. The test uses data for subjects with |R|<b}

\item{tau}{A hypothetical treatment effect. Defaults to 0}

\item{outcome}{A string specifying the name of the outcome variable}

\item{return.coef}{If TRUE return the Z-coefficient from the model; if FALSE return
the p-value}

\item{rhs}{A string specifying the \eqn{\mu_\beta(\cdot)} model relating the outcome to
\code{R}. Defaults to \code{"R+Z"}}
}
\value{
scalar, if return.coef the coefficient on Z; otherwise the p-value associated
w/ coefficient on Z
}
\description{
The test is the test associated with the Z coefficient after fitting
a model from the robustbase suite that uses Z and R as independent variables.
Covariance adjustment is via robust logistic regression if
ytilde is binary, robust linear regression otherwise. In the latter case,
since we're using `robustbase::lmrob` with the `MM` method, the SE used
in the test is of the Huber-White type, allowing for heteroskedasticity
and accounting for  propagation of errors from the preliminary fitting
done by the robust regression routine.
}
\details{
This is the method recommended in version 2 of Limitless
Regression Discontinuity
}
