% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{totalOutcomeSim}
\alias{totalOutcomeSim}
\title{Run the level/power simulation from Table 3}
\usage{
totalOutcomeSim(nreps = 5000)
}
\arguments{
\item{nreps}{Number of simulation replications}
}
\value{
list of output for each simulation run
}
\description{
Run the level/power simulation from Table 3
}
