% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{totalPolySim}
\alias{totalPolySim}
\title{Run the polynomial simulation from Table 4}
\usage{
totalPolySim(nreps = 5000)
}
\arguments{
\item{nreps}{Number of simulation replications}
}
\value{
list of output for each simulation run
}
\description{
Run the polynomial simulation from Table 4
}
