% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_data.R
\docType{data}
\name{wide_data}
\alias{wide_data}
\title{Free Recall Data in Wide Format}
\format{
A data frame of answers for a free recall test data

Sub.ID: the participant id
Response: the response the participant gave to the cue
Disease.Condition: healthy or sick participant condition
}
\usage{
data(wide_data)
}
\description{
Dataset that includes free recall data in long format.
Participants were given a list of words to remember, and
then asked to recall the words. This dataset
is in wide format, which should be converted with arrange
data.
}
\keyword{datasets}
