% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data.R
\name{arrange_data}
\alias{arrange_data}
\title{Arrange Data for Free Recall Scoring}
\usage{
arrange_data(data, responses, sep, id)
}
\arguments{
\item{data}{a dataframe of the variables you would like to return.
Other variables will be included in the returned output in long format.}

\item{responses}{a column name in the dataframe that contains
the participant answers for each item in quotes (i.e., "column")}

\item{sep}{a character separating each response in quotes - example: ",".}

\item{id}{a column name containing participant ID numbers from
the original dataframe}
}
\value{
A dataframe of the participant answers including:
\item{Sub.ID}{The participant id number}
\item{response}{The participant response}
\item{position}{The position number of the response listed}
\item{other}{Any additional columns included}
}
\description{
This function takes wide format free recall data where all
responses are stored in the same cell and converts it to long format.
}
\examples{

#This dataset includes a subject number, set of answers, and
#experiment condition.

load("data/wide_data.Rdata")
DF_long <- arrange_data(
 data = wide_data,
 responses = "Response",
 sep = ",",
 id = "Sub.ID")
head(DF_long)

}
\keyword{arrange,}
\keyword{data,}
\keyword{long}
\keyword{wide,}
