% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metadata.R
\name{load_metadata}
\alias{load_metadata}
\title{Load Metadata}
\usage{
load_metadata(webaddress = "https://rb.gy/l82i9p")
}
\arguments{
\item{webaddress}{The default value for webaddress is the current location
of the metadata list.}
}
\value{
\item{metadata}{The metadata list of avaliable datasets}
}
\description{
This function loads the current metadata avaliable from the
\href{https://github.com/orgs/SemanticPriming/}{Semantic Priming GitHub Group}.
}
\examples{

#Use the following to load the metadata:
metadata <- load_metadata()
head(metadata)
}
\keyword{datasets,}
\keyword{linguistic}
\keyword{metadata,}
\keyword{norms}
